+++ CHANGELOG.md
@@ -12,6 +12,7 @@ _This release is scheduled to be released on 2023-10-01._
 ### Added
 
 - Added UV Index support to OpenWeatherMap
+- Added 'hideDuplicates' flag to the calendar module
 
 ### Removed